# docs/swagger.yaml
basePath: /
definitions:
  requests.CreateSubscriptionRequest:
    properties:
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      price:
        example: 999
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  requests.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      price:
        example: 1299
        type: integer
      service_name:
        example: Netflix Premium
        type: string
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      message:
        example: Detailed error message
        type: string
    type: object
  response.TotalCostResponse:
    properties:
      total:
        example: 15000
        type: integer
    type: object
  test-assignment_internal_app_models.Subscription:
    properties:
      end_date:
        example: "2024-12-31T23:59:59Z"
        type: string
      id:
        example: 1
        type: integer
      price:
        example: 999
        minimum: 1
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
host: localhost:5000
info:
  contact: {}
  description: API для управления подписками пользователей
  title: Subscription Management API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            items:
              $ref: '#/definitions/test-assignment_internal_app_models.Subscription'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/requests.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная подписка
          schema:
            $ref: '#/definitions/test-assignment_internal_app_models.Subscription'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Подписка успешно удалена
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Подписка
          schema:
            $ref: '#/definitions/test-assignment_internal_app_models.Subscription'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
    put:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная подписка
          schema:
            $ref: '#/definitions/test-assignment_internal_app_models.Subscription'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
  /total:
    get:
      description: Вычисляет общую стоимость подписок по заданным критериям за период
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Дата начала периода в формате MM-YYYY
        in: query
        name: from
        required: true
        type: string
      - description: Дата окончания периода в формате MM-YYYY
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Общая стоимость
          schema:
            $ref: '#/definitions/response.TotalCostResponse'
        "400":
          description: Некорректные параметры запроса
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить общую стоимость подписок
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Операции с подписками
  name: subscriptions